update block return after time use.

+++ API/return-cancel.php
@@ -65,27 +65,32 @@
             $action = $data -> action;
             $parking = $data -> parking;
 
-            $sql = "DELETE FROM t_cars WHERE code = '$code' AND datetimeUse = '$datetimeUse';
+            if(new DateTime() > new DateTime($datetimeUse)){
+                $sql = "DELETE FROM t_cars WHERE code = '$code' AND datetimeUse = '$datetimeUse';
                     INSERT INTO t_cars_logger (cars,name,code,agent,tel,datetime,datetimeUse,datetimeReturn,purpose,action,parking) 
                     VALUES ('$cars','$name','$code','$agent','$tel','$datetime','$datetimeUse','$datetimeReturn','$purpose','$action', '$parking');";
 
-            $result = $conn -> query($sql);
+                $result = $conn -> query($sql);
 
-            if($result -> rowCount() > 0){
-                $result -> closeCursor();
-                $msg = msg_line_notify($datetimeUse, $cars, $conn);
-                notify_message($msg, $token);
-                
-                if($action == "returned"){
-                    $msgRetrun ="\n". "ผู้ใช้รถเลขที่ทะเบียน " . $cars . "\n" ."ได้ทำการคืนกุญแจแล้ว" . "\n" . "คิวต่อไปสามารถใช้รถได้แล้วค่ะ.";
-                    notify_message($msgRetrun, $token);
+                if($result -> rowCount() > 0){
+                    $result -> closeCursor();
+                    $msg = msg_line_notify($datetimeUse, $cars, $conn);
+                    notify_message($msg, $token);
+                    
+                    if($action == "returned"){
+                        $msgRetrun ="\n". "ผู้ใช้รถเลขที่ทะเบียน " . $cars . "\n" ."ได้ทำการคืนกุญแจแล้ว" . "\n" . "คิวต่อไปสามารถใช้รถได้แล้วค่ะ.";
+                        notify_message($msgRetrun, $token);
+                    }
+
+                    echo json_encode(['message' => 'Insert Data Complete', 'state' => true]);
+                    // http_response_code(200);
+                }else{
+                    echo json_encode(['message' => 'Error', 'state' => false]);
+                    // http_response_code(400);
                 }
-
-                echo json_encode(['message' => 'Insert Data Complete', 'state' => true]);
-                // http_response_code(200);
             }else{
-                echo json_encode(['message' => 'Error', 'state' => false]);
-                // http_response_code(400);
+                echo json_encode(['message' => 'The car is not in use.', 'state' => false]);
+            // http_response_code(400);
             }
         }else{
             echo json_encode(['message' => 'Error', 'state' => false]);
